vpnbot/requirements.txt
aiogram>=3.17.0
aiohttp==3.10.5
remnawave>=2.0.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
pyyaml==6.0.2
